requirements.txt
spotipy==2.17.1
paho-mqtt>=1.5
